+++ lib/core/data/remote/models/request/referral_status_request.g.dart
@@ -1,0 +1,15 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'referral_status_request.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+ReferralStatusRequest _$ReferralStatusRequestFromJson(
+  Map<String, dynamic> json,
+) => ReferralStatusRequest(phone: json['phone'] as String?);
+
+Map<String, dynamic> _$ReferralStatusRequestToJson(
+  ReferralStatusRequest instance,
+) => <String, dynamic>{'phone': instance.phone};